Fix error on orphaned Edorg when rendering EdOrg Tree for app enable / authorization
edorg_tree_helper.rb raises an exception when there is an orphaned
edorg. This prevents anyone from enabling or authorizing apps in
Admin-Tools when such a situation occours. This fix logs an error to
the rails console, omits the orphaned edorg from the tree and displays
the missing edorgs below the tree.